backend/spartamarket/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autopep8==2.1.0
axios==0.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-seed==0.3.1
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.0
executing==2.0.1
Faker==24.11.0
gprof2dot==2022.7.29
idna==3.7
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
toposort==1.10
traitlets==5.14.2
typing==3.7.4.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
